DatentypÄnderung und Toleranzbereich
Datentyp der x/y-Werte in den Klassen zu double geändert und
Toleranzbereich für die Nullstellen eingerichtet. Prüffunktion
funktioniert, ist aber noch nciht exakt genug

+++ app/src/main/java/com/example/arabellaprivat/tanzderfunktionen/checkAndDraw/Liste.java
@@ -6,7 +6,7 @@ import java.util.ArrayList;
  * Created by Kathi on 10.11.2016.
  */
 
-class Liste extends ArrayList <Float> {
+class Liste extends ArrayList <Double> {
 
 
     /** Constructor */
